+++ src/common/globals.py
