+++ src/genetics/genetics.py
@@ -1,4 +1,7 @@
-from model.model import *
+from model.lesson import Lesson
+from model.schedule import Schedule
+from model.day import Day
+from model.classroom import Classroom
 from common.constants import *
 from common import globals
 from copy import deepcopy
