+++ src/main.py
@@ -1,32 +1,22 @@
-from model.model import *
-from common.constants import *
-from common import globals
-from genetics.genetics import *
+from common.constants import BEST_STREAK_TO_EXIT_ON, GENERATIONS, POPULATION_SIZE
+from common.globals import lesson_count
 
+from model.lesson import Lesson
+from model.schedule import Schedule
+from genetics import genetics
+
+from util import file_utils
 from time import time
 
 PRINT_GENERATIONS = True
-
-def read_file(path: str) -> tuple[list, list[Lesson]]:
-    lessons: list[Lesson] = []
-    with open(path) as f:
-        classrooms: list = f.readline()
-        classrooms = list(map(str.strip, classrooms.split("rooms:")[1].split(",")))
-        f.readline()
-        for line in f.readlines():
-            subject, duration = line.split(",")
-            lessons.append(Lesson(subject.strip(), int(duration)))
-    
-    return classrooms, lessons
-
 
 def print_all_stats(population: list[Schedule]) -> None:
     i: int = 0
     score_sum: int = 0
     same: int = 0
-    score = optimal_criterion(population[0])
+    score = genetics.optimal_criterion(population[0])
     for chromosome in population:
-        prev_score, score = score, optimal_criterion(chromosome)
+        prev_score, score = score, genetics.optimal_criterion(chromosome)
         if score != prev_score:
             if same != 1:
                 print(f"Chromosome:[{i-same}-{i-1}] Score:{score}")
@@ -47,7 +37,7 @@ def print_stats(population: list[Schedule]) -> None:
     i: int = 0
     score_sum: int = 0
     for chromosome in population:
-        score = optimal_criterion(chromosome)
+        score = genetics.optimal_criterion(chromosome)
         score_sum += score
         i += 1
     print(f"Average score: {score_sum / POPULATION_SIZE}")
@@ -56,18 +46,18 @@ def print_stats(population: list[Schedule]) -> None:
 
 def find_best(population: list[Schedule]) -> tuple[Schedule, float]:
     best_schedule = population[0]
-    best_score = optimal_criterion(best_schedule)
+    best_score = genetics.optimal_criterion(best_schedule)
     for chromosome in population:
-        if optimal_criterion(chromosome) > best_score:
+        if genetics.optimal_criterion(chromosome) > best_score:
             best_schedule = chromosome
-            best_score = optimal_criterion(chromosome)
+            best_score = genetics.optimal_criterion(chromosome)
     return best_schedule, best_score
 
 def main() -> None:
-    rooms, lessons = read_file("./data_timetable.txt")
-    globals.lesson_count = len(lessons)
+    rooms, lessons = file_utils.read_file("../data_timetable.txt")
+    lesson_count = len(lessons)
 
-    population: list[Schedule] = generate_population(rooms, lessons)
+    population: list[Schedule] = genetics.generate_population(rooms, lessons)
     
     print("Start population:")
     print_all_stats(population)
@@ -80,10 +70,10 @@ def main() -> None:
     same = 1
 
     for _ in range(GENERATIONS):
-        parents = roulette_selection(population)
-        children = crossover(parents)
-        mutate(children)
-        population = elitism(population, children)
+        parents = genetics.roulette_selection(population)
+        children = genetics.crossover(parents)
+        genetics.mutate(children)
+        population = genetics.elitism(population, children)
 
         if BEST_STREAK_TO_EXIT_ON != -1:
             prev_best_score, best_score = best_score, find_best(population)[1]
